Fixed an issue with division
During the compilation of the result for division the decimal point could be calculated improperly. Due to how the quotient is compiled placing the decimal in the wrong place could create incorrect results.
An example of a calculation that would be incorrect would be `3/40=0.57` instead of `0.075`
This was caused by the decimal point calculation backtracking instead of moving forward in some situations.